vứt session=is admin để test thêm sửa xoá chưa tới màn hình thêm sửa tới được màn hình nhưng ko sửa được xoá được

+++ WebApplication1/Controllers/ShoppingCartController.cs
@@ -157,7 +157,7 @@ namespace WebApplication1.Controllers
                 foreach (var item in cart.Items)
                 {
                     var prodTotal = (item._quantity * item._product.ProductPrice);
-                    var tax = (int)(prodTotal * item._product.Tax);
+                    var tax = prodTotal * (decimal)item._product.Tax;
                     OrderDetail _order_detail = new OrderDetail
                     {
                         ProductID = item._product.ProductID,
